BugFix: 모임 관리 가입 관리 search 에러 헤결
